+++ src/plskern.jl
@@ -96,12 +96,11 @@ res = Jchemo.predict(fm, Xtest; nlv = 1:2)
 res.pred[1]
 res.pred[2]
 
-res = Base.summary(fm, Xtrain, ytrain) ;
+res = summary(fm, Xtrain) ;
 pnames(res)
 z = res.explvarx
-#z = res.explvary
 lines(z.nlv, z.cumpvar,
-    axis = (xlabel = "Nb. LVs", ylabel = "Prop. Explained Variance"))
+    axis = (xlabel = "Nb. LVs", ylabel = "Prop. Explained X-Variance"))
 ```
 """ 
 function plskern(X, Y, weights = ones(nro(X)); nlv, 
@@ -239,11 +238,9 @@ end
 
 """
     summary(object::Plsr, X)
-    summary(object::Plsr, X, Y)
 Summarize the fitted model.
 * `object` : The fitted model.
 * `X` : The X-data that was used to fit the model.
-* `Y` : The Y-data that was used to fit the model.
 """ 
 function Base.summary(object::Plsr, X::Union{Vector, Matrix, DataFrame})
     X = ensure_mat(X)
@@ -261,29 +258,4 @@ function Base.summary(object::Plsr, X::Union{Vector, Matrix, DataFrame})
     (explvarx = explvarx,)
 end
 
-function Base.summary(object::Plsr, X::Union{Vector, Matrix, DataFrame},
-        Y::Union{Vector, Matrix, DataFrame})
-    X = ensure_mat(X)
-    Y = ensure_mat(Y)
-    n, nlv = size(object.T)
-    X = cscale(X, object.xmeans, object.xscales)
-    Y = cscale(Y, object.ymeans, object.yscales)
-    tt = object.TT
-    ## X
-    sstot = sum(object.weights' * (X.^2))
-    tt_adj = vec(sum(object.P.^2, dims = 1)) .* tt
-    pvar = tt_adj / sstot
-    cumpvar = cumsum(pvar)
-    xvar = tt_adj / n    
-    explvarx = DataFrame(nlv = 1:nlv, var = xvar, pvar = pvar, cumpvar = cumpvar)
-    ## y
-    sstot = sum(object.weights' * (Y.^2))
-    tt_adj = vec(sum(object.C.^2, dims = 1)) .* tt
-    pvar = tt_adj / sstot
-    cumpvar = cumsum(pvar)
-    xvar = tt_adj / n    
-    explvary = DataFrame(nlv = 1:nlv, var = xvar, pvar = pvar, cumpvar = cumpvar)
-    ## End
-    (explvarx = explvarx, explvary)
-end